+++ src/main/java/io/github/a5b84/convenientmobgriefing/Mod.java
@@ -1,12 +1,26 @@
 package io.github.a5b84.convenientmobgriefing;
 
+import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.gamerule.v1.GameRuleFactory;
+import net.fabricmc.fabric.api.gamerule.v1.GameRuleRegistry;
+import net.minecraft.world.GameRules;
 import net.minecraft.world.GameRules.BooleanRule;
-import net.minecraft.world.GameRules.Key;
 
-public class Mod {
+public class Mod implements ModInitializer {
 
-    public static Key<BooleanRule> LENIENT_GRIEFING;
-    public static Key<BooleanRule> WITHER_GRIEFING;
-    public static Key<BooleanRule> DRAGON_GRIEFING;
+    public static GameRules.Key<BooleanRule> LENIENT_GRIEFING = register("lenientGriefing");
+    public static GameRules.Key<BooleanRule> WITHER_GRIEFING = register("witherGriefing");
+    public static GameRules.Key<BooleanRule> DRAGON_GRIEFING = register("dragonGriefing");
+
+    /** Enregistre une gamerule */
+    private static GameRules.Key<BooleanRule> register(String name) {
+        return GameRuleRegistry.register(
+                name, GameRules.Category.MOBS,
+                GameRuleFactory.createBooleanRule(true)
+        );
+    }
+
+    @Override
+    public void onInitialize() {}
 
 }